backend/requirements.txt
# FastAPI Server
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
python-multipart>=0.0.9

# MongoDB
pymongo==4.5.0
motor==3.3.1

# CORS
starlette

# HTTP requests
requests>=2.31.0
aiohttp>=3.10.0

# Job Assistant dependencies
beautifulsoup4==4.12.2
openpyxl==3.1.2
aiodns==3.1.1

# Google Gemini for cover letter generation (FREE)
google-generativeai>=0.4.0

# Dev tools (optional)
pytest>=8.0.0
